Fix command execution on windows. (#13)
As discussed on discord the new changes to the lite-xl process api to properly
handle argument escaping on windows introduced on https://github.com/lite-xl/lite-xl/pull/1132,
when passing the command as a table it double quotes all arguments and also
parenthesized statements like `cmd /c (echo hello)` which results in
`"cmd" "/c" "(echo hello)"`. This causes batch to error out as it improperly
interprets "(echo hello)" as a whole command instead of a group of statements.

To fix the issue with newer process api one would need to use a plain
string instead of a table to disable the quoting mechanism when
inappropriate.